+++ packages/image_picker/CHANGELOG.md
@@ -1,6 +1,14 @@
-## 0.5.4+1
+## 0.5.4+3
 
 * Fix the example app failing to load picked video.
+
+## 0.5.4+2
+
+* Request Camera permission if it present in Manifest on Android >= M.
+
+## 0.5.4+1
+
+* Bugfix iOS: Cancel button not visible in gallery, if camera was accessed first.
 
 ## 0.5.4